+++ src/mtool/api/rest/app.py
@@ -37,7 +37,6 @@
 from rest.swordfish.handler import swordfish_api
 #from rest.rest_api.alerts.system_alerts import get_alert_categories_from_influxdb
 from rest.exceptions import InvalidUsage
-from rest.rest_api.volume.volume import create_volume, delete_volume, list_volume, rename_volume, get_max_vol_count, mount_volume, unmount_volume
 #from rest.rest_api.logmanager.logmanager import download_logs
 from util.com.common import get_ip_address, get_hostname
 from rest.rest_api.telemetry.telemetry import set_telemetry_configuration, reset_telemetry_configuration, check_telemetry_endpoint
@@ -60,22 +59,19 @@ import traceback
 import json
 import os
 import eventlet
-import math
 from rest.blueprints.home import home_bp
 from rest.blueprints.pos import pos_bp
 from rest.blueprints.device import device_bp
 from rest.blueprints.array import array_bp
 from rest.blueprints.storage import storage_bp
 from rest.blueprints.disk import disk_bp
+from rest.blueprints.volume import volume_bp
 from rest.db import connection_factory
 from rest.auth import token_required
 from rest.log import logger
 from util.com.common import toJson
 
 eventlet.monkey_patch()
-
-BLOCK_SIZE = 1024 * 1024
-BYTE_FACTOR = 1024
 
 # Serve React App
 app = Flask(__name__, static_folder='./public')
@@ -96,6 +92,7 @@ app.register_blueprint(device_bp)
 app.register_blueprint(array_bp)
 app.register_blueprint(storage_bp)
 app.register_blueprint(disk_bp)
+app.register_blueprint(volume_bp)
 
 @app.after_request
 def after_request(response):
@@ -266,77 +263,6 @@ def delete_subsystem(current_user):
         return abort(404)
 
 
-@app.route('/api/v1/qos', methods=['POST'])
-def qos_create_policies():
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        #vol_names = body.get("volumes")
-        body['vol'] = body.pop('volumes')
-        if "maxbw" in body:
-            body["maxbw"] = int(body["maxbw"])
-        if "minbw" in body:
-            body["minbw"] = int(body["minbw"])
-        if "maxiops" in body:
-            body["maxiops"] = int(body["maxiops"])
-        if "miniops" in body:
-            body["miniops"] = int(body["miniops"])
-        request_body = {
-            "param": body
-        }
-        range_message = "Please enter valid range: \nValid range for IO =  10 - 18446744073709551 KIOPS\nValid range for BW = 10 - 17592186044415 MB/s"
-        response = dagent.qos_create_volume_policies(request_body)
-        response = response.json()
-        if "result" in response and "status" in response["result"] and "code" in response[
-                "result"]["status"] and response["result"]["status"]["code"] == 2080:
-            if "description" in response["result"]["status"]:
-                response["result"]["status"]["description"] += " " + \
-                    range_message
-        return toJson(response)
-    except Exception as e:
-        print("In exception qos_create_policies(): ", e)
-        return make_response('Could not create volume policies', 500)
-
-
-@app.route('/api/v1/qos/reset', methods=['POST'])
-def qos_reset_policies():
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        array_name = body.get("array")
-        vol_names = body.get("volumes")
-        request_body = {
-            "param": {
-                "array": array_name,
-                "vol": vol_names
-            }
-        }
-        response = dagent.qos_reset_volume_policies(request_body)
-        return toJson(response.json())
-    except Exception as e:
-        print("In exception qos_reset_volume_policies(): ", e)
-        return make_response('Could not reset volume policies', 500)
-
-
-@app.route('/api/v1/qos/policies', methods=['POST'])
-def qos_policies():
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        array_name = body.get("array")
-        vol_names = body.get("volumes")
-        request_body = {
-            "param": {
-                "array": array_name,
-                "vol": vol_names
-            }
-        }
-        response = dagent.qos_list_volume_policies(request_body)
-        return toJson(response.json())
-    except Exception as e:
-        print("In exception qos_list_volume_policies(): ", e)
-        return make_response('Could not get volume policies', 500)
-
 
 def validate_email(email):
     regex = "(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])"
@@ -449,24 +375,6 @@ def get_users(current_user):
         return toJson(usersList)
 
 
-"""
-@app.route('/api/v1.0/toggle_status/', methods=['POST'])
-def toggle_status():
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    user_id = body['userid']
-    status = body['status']
-    print(" in app.py toggle_status() ", user_id, status)
-    if user_id != "admin":
-        if connection_factory.toggle_status_from_db(user_id, status):
-            return jsonify({"message": "User status changed"})
-        else:
-            return jsonify({"message": "User status could not changed"})
-    else:
-        return jsonify({"message": "User status changed"})
-"""
-
-
 @app.route('/api/v1.0/update_user/', methods=['POST'])
 def update_user():
     body_unicode = request.data.decode('utf-8')
@@ -551,41 +459,6 @@ def login():
                     'username': received_username})
 
 
-"""
-@app.route('/api/v1.0/bmc_login/', methods=['POST'])
-def bmc_login():
-    body_unicode = request.data.decode('utf-8')
-    #print("came in login")
-    # print(request.authorization)
-
-    body = json.loads(body_unicode)
-    username = body['username']
-    password = body['password']
-    print('username', username, '\n', 'password', password)
-    #col = db['user']
-    # print("col")
-    #user = col.find_one({"_id": re.compile(username, re.IGNORECASE), "password":password, "active": True})
-    # if not user:
-    #    user = col.find_one({"email": re.compile(username, re.IGNORECASE), "password":password, "active": True})
-    # if not user:
-    # print("not user")
-    # return make_response('Could not verify',401)
-    #print('printing the user')
-    #print (user)
-    auth = 'Basic %s' % b64encode(
-        bytes(
-            username + ':' + password,
-            "utf-8")).decode("ascii")
-    print(auth)
-    #token = jwt.encode({'_id':user['_id'], 'exp':datetime.datetime.utcnow() + datetime.timedelta(minutes=60)},app.config['SECRET_KEY'])
-    # print(token)
-    validation = BMC_agent.check_bmc_login(auth)
-    if(validation):
-        return jsonify({'Login': validation, })
-    return jsonify({'Login': validation, }), 401
-"""
-
-
 @app.route('/api/v1.0/get_ip_and_mac', methods=['GET'])
 def getIpAndMac():
     """
@@ -603,217 +476,6 @@ def getIpAndMac():
             return jsonify({"ip": ip, "mac": mac, "host": host, "timestamp": lastUpdatedTime})
     except BaseException:
         return make_response('Could not get ip and mac', 500)
-
-
-# POS expects the volume size to be block aligned
-def make_block_aligned(size):
-    """
-    Example:
-    size = 8.13 GB
-    size = floor( 8.13 * 1000 * 1000 * 1000) =  floor(8130000000) = 8130000000
-    floor(8130000000 / (1024 * 1024)) * (1024 * 1024) = 7753 * (1024 * 1024) = 8129609728 → Pass to POS
-    """
-
-    size = math.floor(size)
-    aligned_size = (math.floor(size / BLOCK_SIZE) * BLOCK_SIZE)
-    return aligned_size
-
-
-@app.route('/api/v1.0/save-volume/', methods=['POST'])
-def saveVolume():
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    # ifOldName=body['old']
-    try:
-        volume_name = body['name']
-        vol_size = float(body['size'])
-        minbw = 0
-        if 'minbw' in body:
-            minbw = body['minbw']
-        maxbw = body['maxbw']
-        array_name = body.get('array')
-        maxiops = int(body['maxiops'])
-        miniops = 0
-        if 'miniops' in body:
-            miniops = body['miniops']
-        unit = body['unit'].upper()
-        mount_vol = body['mount_vol']
-        stop_on_error = body['stop_on_error']
-        count = body['count']
-        subsystem = body['subsystem']
-        suffix = body['suffix']
-        max_available_size = body['max_available_size']
-        iswalvol = False
-        if 'iswalvol' in body:
-            iswalvol = body['iswalvol']
-    except Exception as e:
-        print("Exception Occured in Save Volume")
-        print(e)
-        return make_response('Required fields are missing: ' + e, 400)
-
-    if array_name is None:
-        array_name = dagent.array_names[0]
-
-    if (vol_size == 0):
-        size = int(max_available_size)
-    else:
-        if unit == "MB":
-            size = vol_size * BYTE_FACTOR * BYTE_FACTOR
-        elif unit == "GB":
-            size = vol_size * BYTE_FACTOR * BYTE_FACTOR * BYTE_FACTOR
-        elif unit == 'TB':
-            size = vol_size * BYTE_FACTOR * BYTE_FACTOR * BYTE_FACTOR * BYTE_FACTOR
-        elif unit == 'PB':
-            size = vol_size * BYTE_FACTOR * BYTE_FACTOR * \
-                BYTE_FACTOR * BYTE_FACTOR * BYTE_FACTOR
-        else:
-            # default case assuming the unit is TB
-            size = vol_size * BYTE_FACTOR * BYTE_FACTOR * BYTE_FACTOR * BYTE_FACTOR
-    size = make_block_aligned(size)
-
-    create_vol_res = create_volume(
-        volume_name,
-        array_name,
-        int(size),
-        int(count),
-        int(suffix),
-        mount_vol,
-        stop_on_error,
-        int(maxbw),
-        int(maxiops),
-        int(minbw),
-        int(miniops),
-        subsystem,
-        iswalvol)
-
-    # print("print save volllll", toJson(create_vol_res.json())
-    return toJson(create_vol_res)
-
-
-"""
-
-@app.route('/api/v1.0/save-volume/', methods=['POST'])
-def saveVolume():
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    print("save volumeeeeeeeeeeeeeeee")
-    #ifOldName=body['old']
-    print('----------vol body---', body)
-    volume_name = body['name']
-    vol_size = body['size']
-    maxbw = body['maxbw']
-    maxiops = body['maxiops']
-    description = body['description']
-    unit = body['unit']
-    arrayname = 'ibofArray'
-    mount_vol = body['mount_vol']
-    stop_on_error = body['stop_on_error_checkbox']
-    max_available_size = body['max_available_size']
-    print("vol size",int(vol_size))
-    if (int(vol_size) == 0):
-        print("size is", int(max_available_size))
-        size = int(max_available_size)
-    elif unit== "GB":
-        size=int(vol_size) * 1024 * 1024 * 1024
-    else:
-        size=int(vol_size)* 1024 * 1024 * 1024 * 1024
-    create_vol_res = create_volume(volume_name, int(size), int(maxbw), int(maxiops))
-    #if create_vol_res.status_code != 200:
-    return toJson(create_vol_res.json())
-    #if ("return" in create_vol_res and create_vol_res["return"] == -1) or create_vol_res["result"]["status"]["code"] != "200000":
-    # IP="107.123.123"
-    # portno =0
-    # subnqn="NA"
-    # try:
-    #     col = db['volume']
-    #     volume = col.find_one({"name": volume_name})
-    #     if  not volume:
-    #         agr = [{'$group': { '_id':None,'UsedSpace': {'$sum': '$usedspace'},'TotalSize': {'$sum': '$totalsize'}}}]
-    #         val = list(db.array.aggregate(agr))
-    #         # Arraysize in in pages and each page is 4 kb
-    #         if (val[0]['TotalSize']*4*1024 - val[0]['UsedSpace']) < float(size):
-    #             return abort(404)
-    #         print(size)
-    #         db.array.update_one({"_id":arrayname},{"$set":{"usedspace": val[0]['UsedSpace'] + size}})
-    #         IP = get_ip_address()
-    #         #while(1):
-    #         #    IP=generateRandomIPAddress(4)
-    #         #    found=col.find_one({"ip": IP})
-    #         #    if not found:
-    #         #       break;
-    #         while(1):
-    #             portno = randint(1000, 9999)
-    #             found = col.find_one({"port": portno})
-    #             if not found:
-    #                 break
-    #         #while(1):
-    #         #    subnqn=''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(10))
-    #         #    found = col.find_one({"subnqn": subnqn})
-    #         #    if not found:
-    #         #        break;
-    #         c = db['counters'].find_one({"_id": "volume"})
-    #         c = c['count']
-    #         col.insert_one({'name':volume_name ,'size':vol_size, 'unit':unit,'usedspace':0,'description':description, 'ip':IP ,'port': portno,'subnqn':subnqn,'status':"Active",'accesscount':0,'vol_id':c})
-    #         db.counters.update_one({"_id": "volume"}, {"$inc":{"count": 1}})
-    #     else:
-    #         return abort(404)
-    #     return toJson({'return': 0, 'result': "Volume Created Successfully"})
-    # except:
-    #     print("Operation could not be performed by the db error")
-    #     return abort(404)
-
-"""
-
-'''
-@app.route('/api/v1.0/update-volume/', methods=['PUT'])
-@token_required
-def updateVolume(current_user):
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    update_vol_res = update_volume(body)
-    return toJson(update_vol_res.json())
-'''
-
-
-@app.route('/api/v1.0/volumes/<volume_name>', methods=['PATCH'])
-@token_required
-def renameVolume(current_user, volume_name):
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        if volume_name != body["param"]["name"]:
-            return make_response('URL param not matching with body param', 500)
-    except BaseException:
-        return make_response('Invalid Parameters received', 500)
-    rename_vol_res = rename_volume(body)
-    return toJson(rename_vol_res.json())
-
-
-@app.route('/api/v1.0/volume/mount', methods=['POST'])
-def mountVolume():
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        mount_vol_res = mount_volume(
-            body["name"],
-            body.get("array"),
-            body.get("subnqn"))
-        return toJson(mount_vol_res.json())
-    except Exception as e:
-        print("In exception mountVolume(): ", e)
-        return make_response('Could not mount Volume', 500)
-
-
-@app.route('/api/v1.0/volume/mount', methods=['DELETE'])
-def unmountVolume():
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        unmount_vol_res = unmount_volume(body["name"], body.get("array"))
-        return toJson(unmount_vol_res.json())
-    except Exception as e:
-        print("In exception unmountVolume(): ", e)
-        return make_response('Could not Unmount Volume', 500)
 
 
 @app.route('/api/v1.0/ibofos/mount', methods=['POST'])
@@ -835,115 +497,6 @@ def unmountPOS(current_user):
     except Exception as e:
         print("In exception unmountPOS(): ", e)
         return make_response('Could not unmount POS', 500)
-
-
-@app.route('/api/v1.0/delete_volumes/<name>', methods=['POST'])
-def deleteVolumes(name):
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    vols = list(body['volumes'])
-    total = len(vols)
-    passed = 0
-    fail = 0
-    description = ""
-    print("dellll vol arrayname is", name)
-    # arrays=db.array.find()
-    # arrayname=arrays[0]['_id']
-    # print(arrayname)
-    arrayname = name
-    deleted_vols = []
-    return_msg = {"result": "Success"}
-    for vol in vols:
-        del_vol_cmd = delete_volume(vol, arrayname)
-        if del_vol_cmd.status_code == 200:
-            passed = passed + 1
-            deleted_vols.append(vol)
-            return_msg["result"] = del_vol_cmd.json(
-            )["result"]["status"]["description"]
-
-        elif del_vol_cmd.status_code == 500:
-            fail = fail + 1
-            return_msg["result"] = del_vol_cmd.json()["error"]
-            description += vol["name"] + ": "
-            description += del_vol_cmd.json()["error"]
-            description += "\n"
-
-        else:
-            fail = fail + 1
-            del_vol_cmd = del_vol_cmd.json()
-            if ("result" in del_vol_cmd and "status" in del_vol_cmd["result"]):
-                description += vol["name"] + ":"
-                description += del_vol_cmd["result"]["status"]["description"]
-                description += ", Error code: "
-                description += str(del_vol_cmd["result"]["status"]["code"])
-                description += "\n"
-
-                return_msg["result"] = del_vol_cmd["result"]["status"]
-                return_msg["vol_name"] = vol["name"]
-    # try:
-    #    col = db['volume']
-    #    col.delete_many({'name':{'$in':deleted_vols}})
-    #    usedspace=0
-    #    found = db.volume.count()
-    #    if not found:
-    #        usedspace=0
-    #    else:
-    #        agr = [{'$group': {'_id': None, 'UsedSpace': {'$sum': '$size'}}}]
-    #        val = list(db.volume.aggregate(agr))
-    #        usedspace=val[0]['UsedSpace']
-    #    print(usedspace)
-    #   db.array.update_one({"_id": arrayname}, {"$set": {"usedspace": usedspace}})
-    # except pymongo.error.OperationFailure as e:
-    #    print("Operation could not be performed by the db error", e)
-    if fail == 0:
-        return_msg["return"] = 0
-    else:
-        return_msg["return"] = -1
-
-    return_msg["total"] = total
-    return_msg["passed"] = passed
-    return_msg["failed"] = fail
-    return_msg["description"] = description
-    return toJson(return_msg)
-
-
-@app.route('/api/v1.0/max_volume_count/', methods=['GET'])
-def getMaxVolCount():
-    print("Get max vol count")
-    res = get_max_vol_count()
-    max_vol_count = 256
-
-    if(res.status_code == 200):
-        res = res.json()
-        max_vol_count = res["result"]["data"]["count"]
-    return jsonify(max_vol_count)
-
-
-@app.route('/api/v1/<array_name>/get_volumes/', methods=['GET'])
-def getVolumes(array_name):
-    volumes = list_volume(array_name)
-    for vol in volumes:
-        if "maxiops" in vol:
-            vol["maxiops"] = int(vol["maxiops"])
-    return toJson(volumes)
-
-
-@app.route('/api/v1/get_all_volumes/', methods=['GET'])
-def get_all_volumes():
-    volumes_list = {}
-    try:
-        arrays = dagent.list_arrays()
-        arrays = arrays.json()
-        if "data" in arrays["result"] and "arrayList" in arrays["result"]["data"]:
-            arrays = arrays["result"]["data"]["arrayList"]
-        else:
-            return toJson([])
-        for array in arrays:
-            volumes = list_volume(array["name"])
-            volumes_list[array["name"]] = volumes
-    except Exception as e:
-        print("Exception in /api/v1/get_all_volumes/ ", e)
-    return toJson(volumes_list)
 
 
 @app.route('/api/v1/pos/property', methods=['GET'])
@@ -1047,442 +600,11 @@ def get_hardware_health(current_user):
     except Exception as e:
         return make_response('Could not get hardware health metrics'+str(e), 500)
 
-
-'''
-<pre>{&apos;alertName&apos;: &apos;sdfsdf&apos;, &apos;alertType&apos;: &apos;&apos;, &apos;alertCondition&apos;: &apos;Greater Than&apos;, &apos;alertRange&apos;: &apos;7&apos;, &apos;description&apos;: &apos;sdfgsee eee&apos;}
-</pre>
-'''
-
-'''
-@app.route('/api/v1.0/add_alert/', methods=['POST'])
-@token_required
-def addAlert(current_user):
-    """
-    create an alert through kapacitor and make the entry in database.
-    :param current_user:
-    :return:
-    """
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    #print(body)
-    alertName = body['alertName']
-    alertCondition = body['alertCondition']
-    alertRange = body['alertRange']
-    description = body['description']
-    alertCluster = body['alertCluster']
-
-    alertType = ""
-    alertSubCluster = ""
-
-    alertField = alertFields[alertCluster]
-    result = connection_factory.add_alert_in_db(
-                alertName,
-                alertCluster,
-                alertSubCluster,
-                alertType,
-                alertCondition,
-                alertField,
-                description,
-                alertRange,
-                True)
-    return result
-
-@app.route('/api/v1.0/get_alerts/', methods=['GET'])
-def get_alerts():
-    """
-    fetch the alerts from the DB
-    :return:
-    """
-    alerts = connection_factory.get_alerts_from_db()
-    print("alerts :", alerts)
-    if not alerts:
-        return jsonify({'message': 'no alerts found'})
-    else:
-        return toJson(alerts)
-
-
-@app.route('/api/v1.0/alert', methods=['POST'])
-def alertHandler():
-    #print("alert handler called")
-    #body_unicode = request.data.decode('utf-8')
-    #body = json.loads(body_unicode)
-    allAlerts = get_alerts()
-    # print(list(allAlerts.get_points()))
-    return jsonify(list(allAlerts.get_points()))
-
-
-@app.route('/api/v1.0/delete_alerts/', methods=['POST'])
-@token_required
-def delete_alerts(current_user):
-    """
-    Deletes the given alert from kapacitor and the DB
-    :param current_user:
-    :return: response status
-    """
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    ids = body['ids']
-    for alert_id in ids:
-        result = connection_factory.delete_alerts_in_db(alert_id)
-    return result
-
-
-@app.route('/api/v1.0/update_alerts/', methods=['POST'])
-@token_required
-def update_alerts(current_user):
-    """
-    Api endpoint to update an existing alert.
-    :param current_user:
-    :return: http response with status code
-    """
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    alertName = body['alertName']
-    alertCondition = body['alertCondition']
-    alertRange = body['alertRange']
-    description = body['description']
-    alertCluster = body['alertCluster']
-
-    alertType = ""
-    alertSubCluster = ""
-
-    alertField = alertFields[alertCluster]
-    result = connection_factory.update_alerts_in_db(
-                alertName,
-                alertCluster,
-                alertSubCluster,
-                alertType,
-                alertCondition,
-                alertField,
-                description,
-                alertRange)
-    return result
-
-
-@app.route('/api/v1.0/toggle_alert_status/', methods=['POST'])
-@token_required
-def toggle_alert_status(current_user):
-    """
-    Toggle the state of a given alert
-    :param current_user: token for maintaining the session
-    :return: status wether the alert is successfully changed or not
-    """
-    body_unicode = request.data.decode('utf-8')
-    body = json.loads(body_unicode)
-    print(body)
-    alertName = body['alertName']
-    status = body['status']
-    result = connection_factory.toggle_alert_status_in_db(alertName, status)
-    return result
-
-@app.route('/api/v1.0/download_logs', methods=['GET'])
-def downloadLogs():
-    print(request.args)
-    startdate = request.args.get('start_date')
-    enddate = request.args.get('end_date')
-    return download_logs(startdate, enddate)
-
-
-@app.route('/api/v1.0/set_live_logs/', methods=['POST'])
-@token_required
-def set_live_logs(current_user):
-    """
-    update the live log flag in db
-    :return:
-    """
-    print("set_live_logs palak")
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        print(
-            " in set_live_logs",
-            body,
-            body['liveLogs'],
-            type(
-                body['liveLogs']))
-        connection_factory.set_live_logs_in_db(body['liveLogs'], current_user)
-    except Exception as e:
-        print(e)
-    return "Success"
-
-
-@app.route('/api/v1.0/get_live_logs/', methods=['GET'])
-@token_required
-def get_live_logs(current_user):
-    """
-    update the live log flag in db
-    :return:
-    """
-    print("get_live_logs palak", current_user)
-    livedata = "yes"
-    try:
-        result = connection_factory.get_live_logs_from_db(current_user)
-        print("result ", result, type(result))
-        if result:
-            livedata = "yes"
-        else:
-            livedata = "no"
-    except Exception as e:
-        print(e)
-    print("returning", livedata)
-    return jsonify(livedata)
-'''
-
-
 @app.errorhandler(InvalidUsage)
 def handle_invalid_usage(error):
     response = jsonify(error.to_dict())
     response.status_code = error.status_code
     return response
-
-
-# BMC Rest APIs
-'''
-@app.route('/api/v1.0/get_server_info/', methods=['GET'])
-@token_required
-def getServerInfo(current_user):
-
-    try:
-        return BMC_agent.get_server_info()
-    except BaseException:
-        return make_response('Could not get server info', 500)
-
-
-@app.route('/api/v1.0/get_power_info/', methods=['GET'])
-@token_required
-def getPowerInfo(current_user):
-
-    try:
-        return BMC_agent.get_power_info()
-    except BaseException:
-        return make_response('Could not get power info', 500)
-
-
-@app.route('/api/v1.0/get_chassis_front_info/', methods=['GET'])
-@token_required
-def getChassisFrontInfo(current_user):
-
-    try:
-        return BMC_agent.get_chassis_front_info()
-    except BaseException:
-        return make_response('Could not chassis front info', 500)
-
-"""
-@app.route('/api/v1.0/get_chassis_rear_info/', methods=['GET'])
-@token_required
-def getChassisRearInfo(current_user):
-
-    try:
-        return BMC_agent.get_chassis_rear_info()
-    except BaseException:
-        return make_response('Could not get chassis rear info', 500)
-"""
-
-
-@app.route('/api/v1.0/power_on_system/', methods=['POST'])
-@token_required
-def powerOnSystem(current_user):  # Power On
-
-    try:
-        return BMC_agent.power_on_system()
-    except BaseException:
-        return make_response('Could not power on system', 500)
-
-
-@app.route('/api/v1.0/reboot_system/', methods=['POST'])
-@token_required
-def rebootSystem(current_user):
-    try:
-        return BMC_agent.reboot_system()  # os.system("reboot");
-    except BaseException:
-        return make_response('Could not reboot system', 500)
-
-
-@app.route('/api/v1.0/shutdown_system/', methods=['POST'])
-@token_required
-def shutdownSystem(current_user):
-
-    try:
-        return BMC_agent.shutdown_system()  # os.system("shutdown now -h");
-    except BaseException:
-        return make_response('Could not shutdown system', 500)
-
-@app.route('/api/v1.0/force_shutdown_system/', methods=['POST'])
-@token_required
-def forceShutdownSystem(current_user):
-
-    try:
-        return BMC_agent.force_shutdown_system()
-    except BaseException:
-        return make_response('Could not shutdown system', 500)
-
-
-@app.route('/api/v1.0/get_power_sensor_info/', methods=['GET'])
-@token_required
-def getPowerSensorInfo(current_user):
-
-    try:
-        return BMC_agent.get_power_sensor_info()
-    except BaseException:
-        return make_response('Could not get power sensor info', 500)
-
-@app.route('/api/v1.0/get_fan_sensor_info/', methods=['GET'])
-@token_required
-def getFanSensorInfo(current_user):
-    try:
-        return BMC_agent.get_fan_sensor_info()
-    except BaseException:
-        return make_response('Could not get fan sensor info', 500)
-
-@app.route('/api/v1.0/get_temperature_sensor_info/', methods=['GET'])
-@token_required
-def getTemperatureSensorInfo(current_user):
-    try:
-        return BMC_agent.get_temperature_sensor_info()
-    except BaseException:
-        return make_response('Could not get temperature sensor info', 500)
-
-@app.route('/api/v1.0/get_software_health/', methods=['GET'])
-@token_required
-def getSoftwareHealth(current_user):
-    try:
-        service = [{"mgmt_service": "OK", "data_service": "OK"}]
-        return jsonify({"software_health": service})
-    except BaseException:
-        return make_response('Could not get software health', 500)
-
-@app.route('/api/v1.0/get_hardware_health/', methods=['GET'])
-@token_required
-def getHardwareHealth(current_user):
-    try:
-        cpu, power, fans, memory, temperature = "OK", "OK", "OK", "OK", "OK"
-        res = BMC_agent.get_power_sensor_info()
-        res = res.data
-        res = json.loads(res)
-        power_arr = res['power_sensor_info']
-        for i in range(0, len(power_arr)):
-            if power_arr[i]['Status']['Health'] != "OK":
-                power = "NOT_OK"
-                break
-            else:
-                power = "OK"
-        res = BMC_agent.get_fan_sensor_info()
-        res = res.data
-        res = json.loads(res)
-        fan_arr = res['fan_sensor_info']
-        for i in range(0, len(fan_arr)):
-            if fan_arr[i]['Status']['Health'] != "OK":
-                fans = "NOT_OK"
-                break
-            else:
-                fans = "OK"
-
-        res = BMC_agent.get_temperature_sensor_info()
-        res = res.data
-        res = json.loads(res)
-        temp_arr = res['temperature_sensor_info']
-        for i in range(0, len(temp_arr)):
-            if temp_arr[i]['Status']['Health'] != "OK":
-                temperature = "NOT_OK"
-                break
-            else:
-                temperature = "OK"
-        hardware_health = [{"cpu": cpu,
-                            "memory": memory,
-                            "power": power,
-                            "fans": fans,
-                            "temperature": temperature}]
-        return jsonify({"hardware_health": hardware_health})
-    except BaseException:
-        return make_response('Could not get hardware health', 500)
-
-
-@app.route('/api/v1.0/get_network_health/', methods=['GET'])
-@token_required
-def getNetworkHealth(current_user):
-    try:
-        network = [{"mgmt_network": "OK",
-                    "client_network": "OK",
-                    "storage_fabric": "OK"}]
-        return jsonify({"network_health": network})
-    except BaseException:
-        return make_response('Could not get network health', 500)
-
-
-
-@app.route('/api/v1.0/get_input_power_variation/<time>', methods=['GET'])
-def fetch_power_variation(time):
-    # print(time)
-    if time not in time_groups.keys():
-        raise InvalidUsage(
-            'Use time from 1m,5m,15m,1h,6h,12h,24h,7d,30d',
-            status_code=404)
-    res = get_input_power_variation(time)
-    return jsonify(res)
-
-
-
-@app.route('/api/v1.0/get_power_summary/', methods=['GET'])
-@token_required
-def getPowerSummary(current_user):
-    try:
-        return BMC_agent.getPowerSummary()
-    except BaseException:
-        return make_response('Could not get network health', 500)
-
-
-@app.route('/api/v1.0/set_current_power_mode/', methods=['POST'])
-@token_required
-def setCurrentPowerMode(current_user):
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        # print(body)
-        new_power_mode = body['newpowermode']
-        return BMC_agent.setCurrentPowerMode(new_power_mode)
-        # BMC_agent.shutdown_system()
-    except BaseException:
-        return make_response('Could not update the value', 500)
-
-
-@app.route('/api/v1.0/update_current_power_state/', methods=['POST'])
-@token_required
-def changeCurrentPowerState(current_user):
-    try:
-        body_unicode = request.data.decode('utf-8')
-        body = json.loads(body_unicode)
-        drive_index = body['@odata.id']
-        drive_index = drive_index[-1]
-        power_state = body['PowerState']
-        return BMC_agent.changeCurrentPowerState(
-            int(drive_index), int(power_state))
-    except BaseException:
-        return make_response('Could not update the value', 500)
-# @app.before_first_request
-def activate_bmc_thread():
-    while True:
-        with app.app_context():
-            if Run_Only_Once.OnlyOnceFlag:
-                BMC_agent.get_basic_redfish_url()
-                Run_Only_Once.OnlyOnceFlag = False
-            BMC_agent.fetch_bmc_logs()
-            # BMC_agent.get_power_sensor_info();
-            time.sleep(5)
-
-
-def activate_power_thread():
-    while True:
-        with app.app_context():
-            BMC_agent.get_power_sensor_info()
-            time.sleep(180)
-
-
-# app.host='localhost'
-#app.debug = True
-# app.use_reloader=False
-# app.port=5010
-# app.Threaded=True
-'''
 
 socketIo = SocketIO(app, cors_allowed_origins="*", async_mode='eventlet')
